+++ plugins/plugin-chart-echarts/src/utils/series.ts
@@ -1,0 +1,70 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { DataRecordValue } from '@superset-ui/chart';
+import { EchartsBaseTimeseriesSeries, EchartsTimeseriesDatum } from '../Timeseries/types';
+
+export const extractTimeseriesSeries = (
+  data: EchartsTimeseriesDatum[],
+): Partial<echarts.EChartOption.Series>[] => {
+  const series = [] as EchartsBaseTimeseriesSeries[];
+
+  const keys = data.length > 0 ? Object.keys(data[0]).filter(key => key !== '__timestamp') : [];
+
+  const rawSeries: Record<string, [Date, DataRecordValue][]> = keys.reduce(
+    (obj, key) => ({
+      ...obj,
+      [key]: [],
+    }),
+    {},
+  );
+
+  data.forEach(row => {
+    // eslint-disable-next-line no-underscore-dangle
+    const timestamp = new Date(row.__timestamp);
+    keys.forEach(key => {
+      rawSeries[key].push([timestamp, row[key]]);
+    });
+  });
+
+  Object.entries(rawSeries).forEach(([key, value]) => {
+    series.push({
+      name: key,
+      data: value,
+    });
+  });
+
+  // @ts-ignore
+  return series;
+};
+
+export const extractSeriesBase = (series: echarts.EChartOption.Series[]): number | null => {
+  let minValue: number | null = null;
+  series
+    .filter(row => row !== undefined)
+    .forEach(seriesEntry => {
+      if (seriesEntry?.data) {
+        // @ts-ignore
+        seriesEntry.data.forEach(row => {
+          minValue =
+            minValue === null || minValue === undefined || row[1] < minValue ? row[1] : minValue;
+        });
+      }
+    });
+  return minValue;
+};
